requirements.txt
# Python dependencies for traffic-manager
#
# Install with: pip install -r requirements.txt

# Database - PostgreSQL connection and connection pooling
psycopg2-binary>=2.9.0

# Cache - Redis client for caching
redis>=5.0.0

# Event Streaming - Kafka producer for event publishing
kafka-python>=2.0.2

# Metrics - Prometheus client for metrics collection
prometheus-client>=0.19.0

# API Framework - Flask for REST API endpoints
flask>=3.0.0

# HTTP Client - For load testing scripts
requests>=2.31.0
